GLState: change native config type to intptr_t
Change type of native config to intptr_t so it can safely hold a pointer.

+++ src/gl-state-egl.cpp
@@ -379,7 +379,7 @@ GLStateEGL::swap()
 }
 
 bool
-GLStateEGL::gotNativeConfig(int& vid)
+GLStateEGL::gotNativeConfig(intptr_t& vid)
 {
     if (!gotValidConfig())
         return false;
